# artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ASC Inventory API",
    "version": "1.4.0",
    "description": "Auto-generated from @asc/contract route definitions."
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/catalog": {
      "get": {
        "operationId": "catalog.list",
        "summary": "List all catalog items",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "category",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "includeInactive",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "name": {
                                "type": "string"
                              },
                              "description": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "category": {
                                "type": "string",
                                "enum": [
                                  "IMPLANT",
                                  "INSTRUMENT",
                                  "EQUIPMENT",
                                  "MEDICATION",
                                  "CONSUMABLE",
                                  "PPE"
                                ]
                              },
                              "manufacturer": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "catalogNumber": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "requiresSterility": {
                                "type": "boolean"
                              },
                              "isLoaner": {
                                "type": "boolean"
                              },
                              "active": {
                                "type": "boolean"
                              },
                              "requiresLotTracking": {
                                "type": "boolean"
                              },
                              "requiresSerialTracking": {
                                "type": "boolean"
                              },
                              "requiresExpirationTracking": {
                                "type": "boolean"
                              },
                              "criticality": {
                                "type": "string",
                                "enum": [
                                  "CRITICAL",
                                  "IMPORTANT",
                                  "ROUTINE"
                                ]
                              },
                              "readinessRequired": {
                                "type": "boolean"
                              },
                              "expirationWarningDays": {
                                "type": [
                                  "number",
                                  "null"
                                ]
                              },
                              "substitutable": {
                                "type": "boolean"
                              },
                              "inventoryCount": {
                                "type": "number"
                              },
                              "imageCount": {
                                "type": "number"
                              },
                              "identifierCount": {
                                "type": "number"
                              },
                              "createdAt": {
                                "type": "string"
                              },
                              "updatedAt": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "description",
                              "category",
                              "manufacturer",
                              "catalogNumber",
                              "requiresSterility",
                              "isLoaner",
                              "active",
                              "requiresLotTracking",
                              "requiresSerialTracking",
                              "requiresExpirationTracking",
                              "criticality",
                              "readinessRequired",
                              "expirationWarningDays",
                              "substitutable",
                              "inventoryCount",
                              "imageCount",
                              "identifierCount",
                              "createdAt",
                              "updatedAt"
                            ]
                          }
                        }
                      },
                      "required": [
                        "items"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "catalog.create",
        "summary": "Create a new catalog item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "IMPLANT",
                      "INSTRUMENT",
                      "EQUIPMENT",
                      "MEDICATION",
                      "CONSUMABLE",
                      "PPE"
                    ]
                  },
                  "manufacturer": {
                    "type": "string",
                    "maxLength": 255
                  },
                  "catalogNumber": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "requiresSterility": {
                    "type": "boolean"
                  },
                  "isLoaner": {
                    "type": "boolean"
                  },
                  "requiresLotTracking": {
                    "type": "boolean"
                  },
                  "requiresSerialTracking": {
                    "type": "boolean"
                  },
                  "requiresExpirationTracking": {
                    "type": "boolean"
                  },
                  "criticality": {
                    "type": "string",
                    "enum": [
                      "CRITICAL",
                      "IMPORTANT",
                      "ROUTINE"
                    ]
                  },
                  "readinessRequired": {
                    "type": "boolean"
                  },
                  "expirationWarningDays": {
                    "type": [
                      "integer",
                      "null"
                    ],
                    "exclusiveMinimum": 0
                  },
                  "substitutable": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "name",
                  "category"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "item": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "name": {
                              "type": "string"
                            },
                            "description": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "category": {
                              "type": "string",
                              "enum": [
                                "IMPLANT",
                                "INSTRUMENT",
                                "EQUIPMENT",
                                "MEDICATION",
                                "CONSUMABLE",
                                "PPE"
                              ]
                            },
                            "manufacturer": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "catalogNumber": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "requiresSterility": {
                              "type": "boolean"
                            },
                            "isLoaner": {
                              "type": "boolean"
                            },
                            "active": {
                              "type": "boolean"
                            },
                            "requiresLotTracking": {
                              "type": "boolean"
                            },
                            "requiresSerialTracking": {
                              "type": "boolean"
                            },
                            "requiresExpirationTracking": {
                              "type": "boolean"
                            },
                            "criticality": {
                              "type": "string",
                              "enum": [
                                "CRITICAL",
                                "IMPORTANT",
                                "ROUTINE"
                              ]
                            },
                            "readinessRequired": {
                              "type": "boolean"
                            },
                            "expirationWarningDays": {
                              "type": [
                                "number",
                                "null"
                              ]
                            },
                            "substitutable": {
                              "type": "boolean"
                            },
                            "inventoryCount": {
                              "type": "number"
                            },
                            "imageCount": {
                              "type": "number"
                            },
                            "identifierCount": {
                              "type": "number"
                            },
                            "createdAt": {
                              "type": "string"
                            },
                            "updatedAt": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "description",
                            "category",
                            "manufacturer",
                            "catalogNumber",
                            "requiresSterility",
                            "isLoaner",
                            "active",
                            "requiresLotTracking",
                            "requiresSerialTracking",
                            "requiresExpirationTracking",
                            "criticality",
                            "readinessRequired",
                            "expirationWarningDays",
                            "substitutable",
                            "inventoryCount",
                            "imageCount",
                            "identifierCount",
                            "createdAt",
                            "updatedAt"
                          ]
                        }
                      },
                      "required": [
                        "item"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/catalog/{catalogId}/images": {
      "get": {
        "operationId": "catalog.listImages",
        "summary": "List images for a catalog item",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "catalogId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "images": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "catalogId": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "kind": {
                                "type": "string"
                              },
                              "caption": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "sortOrder": {
                                "type": "number"
                              },
                              "assetUrl": {
                                "type": "string"
                              },
                              "source": {
                                "type": "string"
                              },
                              "createdAt": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "catalogId",
                              "kind",
                              "caption",
                              "sortOrder",
                              "assetUrl",
                              "source",
                              "createdAt"
                            ]
                          }
                        }
                      },
                      "required": [
                        "images"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/catalog/{catalogId}/images/{imageId}": {
      "delete": {
        "operationId": "catalog.deleteImage",
        "summary": "Delete a catalog image",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "catalogId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "imageId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          }
        }
      }
    }
  },
...
}